max message handling uses counting bloom filter

+++ bindings/python_cffi/include/neuropil_comb.h
@@ -77,8 +77,8 @@ struct np_settings {
   uint32_t              log_level;
   uint8_t               leafset_size;
   np_log_write_callback log_write_fn;
-  size_t                jobqueue_size;
-  size_t                max_msgs_per_sec;
+  uint16_t              jobqueue_size;
+  uint16_t              max_msgs_per_sec;
 };
 struct np_settings *np_default_settings(struct np_settings *settings);
 np_context         *np_new_context(struct np_settings *settings);
